+++ src/routes/me/index.ts
@@ -1,19 +1,21 @@
-import status from "./session";
+import session from "./session";
 import blocked from "./blocked";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 import votes from "./votes";
 import starred from "./starred";
 import sessions from "./sessions";
+import logout from "./logout";
 
 export default function (
     fastify: FastifyInstance,
     _opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) {
-    fastify.register(status);
+    fastify.register(logout);
+    fastify.register(session);
+    fastify.register(sessions, { prefix: "/sessions" });
+    fastify.register(votes, { prefix: "/votes" });
     fastify.register(blocked);
     fastify.register(starred);
-    fastify.register(sessions);
-    fastify.register(votes, { prefix: "/votes" });
     done();
 }